+++ pkg/models/transaction/transaction.go
@@ -1,4 +1,4 @@
-package meta_transaction
+package transaction
 
 import (
 	"context"
